requirements.txt
kamidana
slack_sdk
python-dotenv
